+++ src/statistics.rs
@@ -24,7 +24,7 @@ pub fn cardinality(sketch: &[KmerCount]) -> Result<u64, &'static str> {
 ///
 pub fn hist(sketch: &[KmerCount]) -> Vec<u64> {
     let mut counts = vec![0u64; 65536];
-    let mut max_count: u16 = 0;
+    let mut max_count: u64 = 0;
     for kmer in sketch {
         max_count = cmp::max(max_count, kmer.count);
         counts[kmer.count as usize - 1] += 1;
